Full read-over and revision.
Removed a section ... it now seems extraneous. It was the section linking to all sorts of things from Gideon Wulfsohn's document. That document is still linked-to at bottom of guide. Thing is, Wulfsohn revised his document too! He removed some stuff. Seems like I should too. Just the kind of thought that comes with revision I guess. You realize some things are contradictory. Putting in this "list of lots of cool things" countradicts some other arguments in the guide (i.e. _the algorithm may not be as important as the data_).

Then I got on a roll. Kept reading and made various revisions.