Rewrote the section about ML and UX
It's an important point but the earlier version of this section was poorly written (my bad!). I have replaced it with a new take that is more general and points out to other resources moreso. Hopefully good enough.

Also drilled in the more general point that working with problem domain experts will yield better results.