Small update re: UML and ESL books
The old link to ESL was broken, but I found a new working link (which also _appears_ to be legit, it is hosted on stanford.edu after all, under `~hastie` and Hastie is one of the authors). Also rephrased that section a little - this seems clearer and more useful to the reader.